requirements.txt
# SR DataLake (SDL) - Python Dependencies
flask==3.1.2
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pymysql==1.1.2
minio==7.2.20
paho-mqtt==2.1.0
requests==2.32.5
bcrypt==5.0.0
paramiko==4.0.0
psutil==7.2.2
pymodbus==3.12.0
lxml==6.0.2
python-docx==1.2.0
gunicorn==23.0.0
python-dotenv==1.1.1
